+++ packages/stream_video_flutter/lib/src/call_screen/call_content/call_content.dart
@@ -51,7 +51,7 @@ class StreamCallContent extends StatefulWidget {
     this.callControlsWidgetBuilder,
     this.callNotConnectedBuilder,
     this.callFastReconnectingOverlayBuilder,
-    this.layoutMode = ParticipantLayoutMode.grid,
+    this.layoutMode = ParticipantLayoutMode.auto,
     this.extendBody = false,
     this.pictureInPictureConfiguration = const PictureInPictureConfiguration(),
   });